Rename 'update' to 'upgrade'
