+++ application/src/main/java/org/thingsboard/server/vsensor/upgrade/component/RateLimitsUpdater.java
@@ -16,7 +16,7 @@
 /*
 * Ahmet Ertuğrul KAYA
 */
-package org.thingsboard.server.vsensor.update.component;
+package org.thingsboard.server.vsensor.upgrade.component;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
