+++ application/src/main/java/org/thingsboard/server/vsensor/upgrade/configuration/DeviceConnectivityUpdateConfiguration.java
@@ -16,7 +16,7 @@
 /*
 * Ahmet Ertuğrul KAYA
 */
-package org.thingsboard.server.vsensor.update.configuration;
+package org.thingsboard.server.vsensor.upgrade.configuration;
 
 import lombok.Data;
 import org.springframework.boot.context.properties.ConfigurationProperties;
@@ -27,11 +27,11 @@ import org.thingsboard.server.dao.device.DeviceConnectivityInfo;
 import java.util.HashMap;
 import java.util.Map;
 
-@Profile("update")
+@Profile("upgrade")
 @Configuration
 @ConfigurationProperties(prefix = "device")
 @Data
-public class DeviceConnectivityConfiguration {
+public class DeviceConnectivityUpdateConfiguration {
     private Map<String, DeviceConnectivityInfo> connectivity = new HashMap<>();
 
     public boolean isEnabled(String protocol) {
