+++ application/src/main/java/org/thingsboard/server/vsensor/upgrade/service/DatabaseUpgradeService.java
@@ -16,22 +16,10 @@
 /*
 * Ahmet Ertuğrul KAYA
 */
-package org.thingsboard.server.vsensor.update.service;
+package org.thingsboard.server.vsensor.upgrade.service;
 
-import org.thingsboard.server.common.data.security.model.JwtSettings;
+public interface DatabaseUpgradeService {
+    void upgradeDatabase(String fromVersion) throws Exception;
 
-public interface LegacyJwtSettingsService {
-
-    String ADMIN_SETTINGS_JWT_KEY = "jwt";
-    String TOKEN_SIGNING_KEY_DEFAULT = "thingsboardDefaultSigningKey";
-
-    void createRandomJwtSettings();
-
-    void saveLegacyYmlSettings();
-
-    JwtSettings getJwtSettings();
-
-    JwtSettings reloadJwtSettings();
-
-    JwtSettings saveJwtSettings(JwtSettings jwtSettings);
+    String getCurrentSchemeVersion() throws Exception;
 }
