+++ application/src/main/java/org/thingsboard/server/vsensor/upgrade/service/DbUpgradeExecutorService.java
@@ -16,26 +16,14 @@
 /*
 * Ahmet Ertuğrul KAYA
 */
-package org.thingsboard.server.vsensor.update.configuration;
+package org.thingsboard.server.vsensor.upgrade.service;
 
-import lombok.Data;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
-import org.thingsboard.server.dao.device.DeviceConnectivityInfo;
+import org.springframework.stereotype.Component;
+import org.thingsboard.server.service.executors.DbCallbackExecutorService;
 
-import java.util.HashMap;
-import java.util.Map;
+@Component
+@Profile("upgrade")
+public class DbUpgradeExecutorService extends DbCallbackExecutorService {
 
-@Profile("update")
-@Configuration
-@ConfigurationProperties(prefix = "device")
-@Data
-public class DeviceConnectivityUpdateConfiguration {
-    private Map<String, DeviceConnectivityInfo> connectivity = new HashMap<>();
-
-    public boolean isEnabled(String protocol) {
-        var info = connectivity.get(protocol);
-        return info != null && info.isEnabled();
-    }
 }
