+++ application/src/main/java/org/thingsboard/server/vsensor/upgrade/service/DefaultSystemUpgradeDataLoaderService.java
@@ -16,7 +16,7 @@
 /*
 * Ahmet Ertuğrul KAYA
 */
-package org.thingsboard.server.vsensor.update.service;
+package org.thingsboard.server.vsensor.upgrade.service;
 
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.google.common.util.concurrent.FutureCallback;
@@ -73,7 +73,7 @@ import org.thingsboard.server.dao.user.UserService;
 import org.thingsboard.server.service.install.InstallScripts;
 import org.thingsboard.server.service.install.SystemDataLoaderService;
 import org.thingsboard.server.service.security.auth.jwt.settings.JwtSettingsService;
-import org.thingsboard.server.vsensor.update.configuration.DeviceConnectivityUpdateConfiguration;
+import org.thingsboard.server.vsensor.upgrade.configuration.DeviceConnectivityUpdateConfiguration;
 
 import javax.annotation.Nullable;
 import javax.annotation.PostConstruct;
@@ -94,9 +94,9 @@ import static org.thingsboard.server.service.security.auth.jwt.settings.DefaultJ
 import static org.thingsboard.server.service.security.auth.jwt.settings.DefaultJwtSettingsService.validateTokenSigningKeyLength;
 
 @Service
-@Profile("update")
+@Profile("upgrade")
 @Slf4j
-public class DefaultSystemUpdateDataLoaderService implements SystemDataLoaderService {
+public class DefaultSystemUpgradeDataLoaderService implements SystemDataLoaderService {
 
     public static final String CUSTOMER_CRED = "customer";
     public static final String ACTIVITY_STATE = "active";
